+++ src/mindbridge/_telemetry.py
@@ -25,6 +25,7 @@ MODEL_TTFT = "mindbridge.model.time_to_first_token"
 GEN_AI_TTFC = "gen_ai.response.time_to_first_chunk"
 GEN_AI_FINISH_REASONS = "gen_ai.response.finish_reasons"
 EMBEDDING_PARTS_ELIDED = "mindbridge.embedding.elided_parts"
+EMBEDDING_VIDEO_SAMPLED = "mindbridge.embedding.video_sampled_inputs"
 GROUNDING_MEDIA_ELIDED = "mindbridge.grounding.media_elided_hits"
 GROUNDING_HITS_DROPPED = "mindbridge.grounding.dropped_hits"